incorporate PR suggestions

+++ pages/api/snippets/[id]/index.js
@@ -17,15 +17,25 @@ export default async function handler(request, response) {
       return response.status(200).json(snippet);
     }
     if (request.method === "PUT") {
-      const snippetData = request.body;
-      await Snippet.findByIdAndUpdate(id, snippetData);
+      const snippet = await Snippet.findById(id);
+      if (!snippet) {
+        return response.status(404).json({ status: "Snippet not found 😔" });
+      } else {
+        const snippetData = request.body;
+        await Snippet.findByIdAndUpdate(id, snippetData);
 
-      response.status(200).json({ status: "Snippet updated!" });
+        return response.status(200).json({ status: "Snippet updated!" });
+      }
     }
 
     if (request.method === "DELETE") {
-      await Snippet.findByIdAndDelete(id);
-      response.status(200).json({ status: "Snippet deleted" });
+      const snippet = await Snippet.findById(id);
+      if (!snippet) {
+        return response.status(404).json({ status: "Snippet not found 😔" });
+      } else {
+        await Snippet.findByIdAndDelete(id);
+        return response.status(200).json({ status: "Snippet deleted" });
+      }
     }
   } catch (error) {
     console.error("An error occurred:", error);
